Unify dictionary indexing triggers

+++ functions/src/index.ts
@@ -9,19 +9,6 @@ const index = algoliasearch(
 )
     .initIndex("dictionary");
 
-// eslint-disable-next-line require-jsdoc
-function applyIndexing(
-    change: functions.firestore.QueryDocumentSnapshot,
-    context: functions.EventContext
-) {
-  const entry = change.data();
-  const base = {
-    entryID: context.params.entryID,
-    language: context.params.language,
-    forms: entry.forms.map(({plain}: never) => plain),
-    headword: entry.forms[0].plain,
-  };
-
   type Record = {
     entryID: string;
     language: string;
@@ -31,44 +18,42 @@ function applyIndexing(
     definition: string | undefined;
     tags: string[] | undefined;
   };
-  const records = [];
 
-  for (const use of entry.uses) {
-    const record = Object.assign({term: use.term}, base) as Record;
-    const tags = [].concat(...(entry.tags ?? []), ...(use.tags ?? []));
-    if (tags?.length) {
-      record.tags = tags.map((t) => "#"+t);
-    }
-    if (use.definition) {
-      record.definition = use.definition;
-    }
-    records.push(record);
-  }
-
-  return index.saveObjects(
-      records,
-      {autoGenerateObjectIDIfNotExist: true}
-  );
-}
-
-export const addToIndex = functions
+export const indexDictionary = functions
     .region("europe-central2")
     .firestore.document("languages/{language}/dictionary/{entryID}")
-    .onCreate(applyIndexing);
+    .onWrite(async (change, context) => {
+      if (change.before.exists) {
+        await index.deleteBy({
+          filters: "entryID:" + context.params.entryID,
+        });
+      }
 
-export const updateIndex = functions
-    .region("europe-central2")
-    .firestore.document("languages/{language}/dictionary/{entryID}")
-    .onUpdate(async (change, context) => {
-      await index.deleteBy({
-        filters: "entryID:" + context.params.entryID,
-      });
-      applyIndexing(change.after, context);
+      const entry = change.after.data();
+      if (!entry) return;
+
+      const base = {
+        entryID: context.params.entryID,
+        language: context.params.language,
+        forms: entry.forms.map(({plain}: never) => plain),
+        headword: entry.forms[0].plain,
+      };
+      const records = [];
+
+      for (const use of entry.uses) {
+        const record = Object.assign({term: use.term}, base) as Record;
+        const tags = [].concat(...(entry.tags ?? []), ...(use.tags ?? []));
+        if (tags?.length) {
+          record.tags = tags.map((t) => "#"+t);
+        }
+        if (use.definition) {
+          record.definition = use.definition;
+        }
+        records.push(record);
+      }
+
+      await index.saveObjects(
+          records,
+          {autoGenerateObjectIDIfNotExist: true}
+      );
     });
-
-export const deleteFromIndex = functions
-    .region("europe-central2")
-    .firestore.document("languages/{language}/dictionary/{entryID}")
-    .onDelete((_, context) => index.deleteBy({
-      filters: "entryID:" + context.params.entryID,
-    }));
